OPHJOD-153: Move counter example for better structure
