+++ src/state/store.ts
@@ -1,7 +1,7 @@
 import { configureStore } from '@reduxjs/toolkit';
 import { persistStore, persistReducer, FLUSH, REHYDRATE, PAUSE, PERSIST, PURGE, REGISTER } from 'redux-persist';
 import storage from 'redux-persist/lib/storage';
-import counterReducer from '../routes/Home/counterSlice';
+import counterReducer from '../features/counter/counterSlice';
 
 const persistConfig = {
   key: 'root',
